+++ tests/e2e/modules/calendar_spec.js
@@ -1,5 +1,6 @@
 const helpers = require("../global-setup");
 const serverBasicAuth = require("../../servers/basic-auth.js");
+const expect = require("chai").expect;
 
 const describe = global.describe;
 const it = global.it;
@@ -31,8 +32,47 @@ describe("Calendar module", function () {
 			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/default.js";
 		});
 
-		it("Should return TestEvents", function () {
-			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
+		it("should show the default maximumEntries of 10", async () => {
+			await app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
+			const events = await app.client.$$(".calendar .event");
+			return expect(events.length).equals(10);
+		});
+
+		it("should show the default calendar symbol in each event", async () => {
+			await app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
+			const icons = await app.client.$$(".calendar .event .fa-calendar");
+			return expect(icons.length).not.equals(0);
+		});
+	});
+
+	describe("Custom configuration", function () {
+		before(function () {
+			// Set config sample for use in test
+			process.env.MM_CONFIG_FILE = "tests/configs/modules/calendar/custom.js";
+		});
+
+		it("should show the custom maximumEntries of 4", async () => {
+			await app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
+			const events = await app.client.$$(".calendar .event");
+			return expect(events.length).equals(4);
+		});
+
+		it("should show the custom calendar symbol in each event", async () => {
+			await app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
+			const icons = await app.client.$$(".calendar .event .fa-birthday-cake");
+			return expect(icons.length).equals(4);
+		});
+
+		it("should show two custom icons for repeating events", async () => {
+			await app.client.waitUntilTextExists(".calendar", "TestEventRepeat", 10000);
+			const icons = await app.client.$$(".calendar .event .fa-undo");
+			return expect(icons.length).equals(2);
+		});
+
+		it("should show two custom icons for day events", async () => {
+			await app.client.waitUntilTextExists(".calendar", "TestEventDay", 10000);
+			const icons = await app.client.$$(".calendar .event .fa-calendar-day");
+			return expect(icons.length).equals(2);
 		});
 	});
 
@@ -47,7 +87,7 @@ describe("Calendar module", function () {
 			serverBasicAuth.close(done());
 		});
 
-		it("Should return TestEvents", function () {
+		it("should return TestEvents", function () {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
@@ -63,7 +103,7 @@ describe("Calendar module", function () {
 			serverBasicAuth.close(done());
 		});
 
-		it("Should return TestEvents", function () {
+		it("should return TestEvents", function () {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
@@ -79,7 +119,7 @@ describe("Calendar module", function () {
 			serverBasicAuth.close(done());
 		});
 
-		it("Should return TestEvents", function () {
+		it("should return TestEvents", function () {
 			return app.client.waitUntilTextExists(".calendar", "TestEvent", 10000);
 		});
 	});
@@ -95,7 +135,7 @@ describe("Calendar module", function () {
 			serverBasicAuth.close(done());
 		});
 
-		it("Should return No upcoming events", function () {
+		it("should return No upcoming events", function () {
 			return app.client.waitUntilTextExists(".calendar", "No upcoming events.", 10000);
 		});
 	});